Merge github.com:illumos/illumos-gate

Conflicts:
	usr/src/cmd/isainfo/isainfo.c
	usr/src/uts/common/Makefile.files
	usr/src/uts/common/fs/zfs/dmu.c
	usr/src/uts/common/fs/zfs/zvol.c
	usr/src/uts/common/io/mr_sas/fusion.h
	usr/src/uts/common/io/mr_sas/ld_pd_map.c
	usr/src/uts/common/io/mr_sas/ld_pd_map.h
	usr/src/uts/common/io/mr_sas/mr_sas.c
	usr/src/uts/common/io/mr_sas/mr_sas.h
	usr/src/uts/common/io/mr_sas/mr_sas_list.c
	usr/src/uts/common/io/mr_sas/mr_sas_list.h
	usr/src/uts/common/io/mr_sas/mr_sas_tbolt.c
	usr/src/uts/i86pc/io/apix/apix.c
	usr/src/uts/i86pc/os/intr.c
	usr/src/uts/intel/sys/x86_archext.h

+++ usr/src/lib/libinstzones/common/zones_locks.c
@@ -24,7 +24,9 @@
  * Use is subject to license terms.
  */
 
-
+/*
+ * Copyright (c) 2012 by Delphix. All rights reserved.
+ */
 
 /*
  * Module:	zones_locks.c
@@ -208,9 +210,9 @@ _z_acquire_lock(char **r_lockKey, char *a_zoneName, char *a_lockObject,
 
 	if (a_pid > 0) {
 		(void) _z_add_arg(args, "-p");	/* lock valid process i.d. */
-		(void) _z_add_arg(args, "%ld", getpid());
+		(void) _z_add_arg(args, "%lld", (long long)getpid());
 		(void) _z_add_arg(args, "-z");	/* lock valid zone i.d. */
-		(void) _z_add_arg(args, "%ld", getzoneid());
+		(void) _z_add_arg(args, "%lld", (long long)getzoneid());
 	}
 
 	/* execute command */
